rename bits/src "load" -> "exports" to clarify purpose
