+++ ui/bits/src/bits.cropDialog.ts
@@ -3,9 +3,9 @@ import { domDialog } from 'common/dialog';
 import Cropper from 'cropperjs';
 
 export interface CropOpts {
-  aspectRatio: number; // required
+  aspectRatio?: number;
   source?: Blob | string; // image or url
-  max: { megabytes: number; pixels?: number }; // constrain size
+  max?: { megabytes?: number; pixels?: number }; // constrain size
   post?: { url: string; field?: string }; // multipart post form url and field name
   onCropped?: (result: Blob | boolean, error?: string) => void; // result callback
 }
@@ -66,9 +66,9 @@ export async function initModule(o?: CropOpts) {
     class: 'crop-viewer',
     css: [{ hashed: 'bits.cropDialog' }, { url: 'npm/cropper.min.css' }],
     htmlText: `<h2>Crop image to desired shape</h2>
-<div class="crop-view"></div>
-<span class="dialog-actions"><button class="button button-empty cancel">cancel</button>
-<button class="button submit">submit</button></span>`,
+      <div class="crop-view"></div>
+      <span class="dialog-actions"><button class="button button-empty cancel">cancel</button>
+      <button class="button submit">submit</button></span>`,
     append: [{ where: '.crop-view', node: container }],
     actions: [
       { selector: '.dialog-actions > .cancel', listener: (_, d) => d.close() },
@@ -95,7 +95,7 @@ export async function initModule(o?: CropOpts) {
     const tryQuality = (quality = 0.9) => {
       canvas.toBlob(
         blob => {
-          if (blob && blob.size < opts.max.megabytes * 1024 * 1024) submit(blob);
+          if (blob && blob.size < (opts.max?.megabytes ?? 100) * 1024 * 1024) submit(blob);
           else if (blob && quality > 0.05) tryQuality(quality * 0.9);
           else submit(false, 'Rendering failed');
         },
