+++ ui/bits/src/bits.dailyFeed.ts
@@ -1,4 +1,4 @@
-import flairPickerLoader from 'common/flairPicker';
+import flairPickerLoader from './exports/flairPicker';
 
 $('.emoji-details').each(function (this: HTMLElement) {
   flairPickerLoader(this);
