+++ ui/bits/tsconfig.json
@@ -7,6 +7,7 @@
     "skipLibCheck": true,
     "noImplicitReturns": false,
     "declaration": true,
+    "composite": true,
     "allowJs": true
   },
   "references": [
